refactor(cef): extract pane_pool.rs from window_pool.rs
Tier 2 item from docs/reports/REPORT_LARGE_FILE_MODULARIZATION_SCAN_2026_07_22.md.
window_pool.rs bundled two structurally parallel, independently-lifecycled
subsystems already self-labeled in the source: the top-level window pool
(?pool=1 full windows) and the pane pool (floating-pool-* frameless panes,
already commented as a distinct section). Extracted the pane-pool cluster
(HWND cache + spawn/register/promote/orphan-cleanup family) wholesale into
commands/pane_pool.rs. window_pool.rs re-exports it (pub use super::
pane_pool::*), so the 5 external call sites (floating_pane.rs,
memory_heartbeat.rs, client/lifecycle.rs, commands/drag.rs, commands/
floating_pane.rs) needed zero changes.

window_pool.rs: 2256 -> 1688 lines. Did not attempt the further top-level-
pool 3-way split (optional/lower-priority per the audit; the ~460-line
Windows-only promote_pool_window has heavy cfg(windows) asymmetry that
would need more care). cargo build -p agentmux-cef clean; cargo test -p
agentmux-cef window_pool:: — 6/6 pass, same as before.

<!-- agentmux:agent_id=agentx -->
